Move our Atomos boilerplate into AtomosFrontend.
